==> analysis/1_fragsys_input_tables.ipynb <==
# GENERATING MAIN OUTPUT TABLES

This notebook contains the code used to generate the main output tables from the 1,000s of files that FRAGSYS generates. Due to a size constraint, these files are not in the repository. However, the main tables generated here are, and are the basis of the analysis.

## IMPORTING NECESSARY PACKAGES

from fragsys_analysis import *

## READING INPUT DATA

main_dir = "/cluster/gjb_lab/2394007/pandda_analysis/phase4/prots"
prot_dirs = sorted(Path(main_dir).iterdir(), key = os.path.getmtime)
results_dir = "./../results/"

## PROTEIN SUMMARY DATAFRAME

results_dfs = []
for prot_dir in prot_dirs:
    prot = str(prot_dir).split("/")[-1]
    subdirs = sorted(os.listdir(os.path.join(prot_dir, "unsupp_cifs"))) # one for each group
    for subdir in subdirs:
        res_path = os.path.join(prot_dir, "{}_{}_results.csv".format(prot, subdir))
        if os.path.isfile(res_path):
            prot_res = pd.read_csv(res_path)
            results_dfs.append(prot_res)
        else:
            print("Group {} of {} did not present results".format(subdir, prot))
            pass
        
fragsys_results_df = pd.concat(results_dfs).reset_index(drop = True)
fragsys_results_df["vars_per_seq"] = round(fragsys_results_df.n_vars/fragsys_results_df.n_human_seqs, 2)
fragsys_results_df["vars_per_res"] = round(fragsys_results_df.n_vars/fragsys_results_df.n_human_res, 2)

fragsys_results_df.shape # total number of protein segments

fragsys_results_df.head(3)

fragsys_results_df.to_pickle(os.path.join(results_dir, "all_prots.pkl")) # all protein segments in data set

## BINDING SITES DATAFRAME

bs_dfs = []
for prot_dir in prot_dirs:
    prot = str(prot_dir).split("/")[-1]
    subdirs = sorted(os.listdir(os.path.join(prot_dir, "unsupp_cifs")))
    for subdir in subdirs:
        bs_res_path = os.path.join(prot_dir, "results", "{}_{}_BS_df.csv".format(prot, subdir))
        if os.path.isfile(bs_res_path):
            prot_bs_res = pd.read_csv(bs_res_path)
            prot_bs_def = pd.read_csv(os.path.join(prot_dir, "results", "{}_{}_BS_def_OC_single_i_rel_0.66.csv".format(prot, subdir)))
            bs_lig_occ_dict = {"BS"+str(k):v for k, v in prot_bs_def.binding_site.value_counts().to_dict().items()}
            prot_bs_res["number_ligs"] = prot_bs_res.bs_id.map(bs_lig_occ_dict)
            prot_bs_res["prop_ligs"] = prot_bs_res.number_ligs/prot_bs_res.number_ligs.sum()
            prot_bs_res["protein"] = prot
            prot_bs_res["group"] = subdir
            bs_dfs.append(prot_bs_res)
        else:
            print("Group {} of {} did not present binding sites data".format(prot, subdir))
            pass
all_bss_dfs = round(pd.concat(bs_dfs).reset_index(drop = True), 4)

all_bss_dfs.shape # total number of residues with structural information

all_bss_dfs.head(3)

all_bss_dfs.to_pickle(os.path.join(results_dir, "all_bss.pkl")) # all binding sites

## BINDING RESIDUES DATAFRAME

fragsys_dfs = []
for prot_dir in prot_dirs:
    prot = str(prot_dir).split("/")[-1]
    subdirs = sorted(os.listdir(os.path.join(prot_dir, "unsupp_cifs")))
    for subdir in subdirs:
        res_path = os.path.join(prot_dir, "results", "{}_{}_fragsys_df.csv".format(prot, subdir))
        if os.path.isfile(res_path):
            prot_res = pd.read_csv(res_path)
            prot_res["protein"] = prot
            prot_res["group"] = subdir
            fragsys_dfs.append(prot_res)
        else:
            print("Group {} of {} did not present results".format(prot, subdir))
            pass
        
fragsyss_df = pd.concat(fragsys_dfs).reset_index(drop = True)
fragsyss_df.SS = fragsyss_df.SS.fillna("C")

fragsyss_df.shape # total number of ligand binding residues

fragsyss_df.head(3)

fragsyss_df.to_pickle(os.path.join(results_dir, "all_bs_ress.pkl")) # all binding site residues

## DSSP DATAFRAME

dssp_dfs = []
for prot_dir in prot_dirs:
    prot = str(prot_dir).split("/")[-1]
    subdirs = sorted(os.listdir(os.path.join(prot_dir, "unsupp_cifs")))
    for subdir in subdirs:
        dssp_subdir = os.path.join(prot_dir, "results", "dssp", subdir)
        sifts_subdir = os.path.join(prot_dir, "results", "sifts", subdir)
        dssp_files = [file for file in os.listdir(dssp_subdir) if file.startswith("dssp")]
        pdb_id = dssp_files[0].split("_")[1]
        example_dssp = os.path.join(dssp_subdir, dssp_files[0]) #only grabbing one example dssp file
        example_sifts = os.path.join(sifts_subdir, "sifts_mapping_{}.csv".format(pdb_id))
        if os.path.isfile(example_dssp):
            dssp_df = pd.read_csv(example_dssp)
            sifts_df = pd.read_csv(example_sifts)
            dssp_df["protein"] = prot
            dssp_df["group"] = subdir
            dssp_df["structure"] = pdb_id
            merged_df = dssp_df.merge(sifts_df[["PDB_ResNum", "PDB_ChainID", "UniProt_ResNum"]], how = "left", left_on = ["PDB_ResNum", "CHAIN"], right_on = ["PDB_ResNum", "PDB_ChainID"])
            dssp_dfs.append(merged_df)
        else:
            print("Group {} of {} did not present DSSP data".format(prot, subdir))
            pass
        
all_dssp_dfs = pd.concat(dssp_dfs)
all_dssp_dfs_filt = all_dssp_dfs.copy().query('UniProt_ResNum == UniProt_ResNum')
all_dssp_dfs_filt.UniProt_ResNum = all_dssp_dfs_filt.UniProt_ResNum.astype(int)

all_dssp_dfs_filt.shape # total number of residues with structural information

all_dssp_dfs_filt.head(3)

all_dssp_dfs_filt.to_pickle(os.path.join(results_dir, "all_dssp_dfs.pkl")) # all dssp data frames

## VARIANTS DATAFRAME

missense_dfs = []
for prot_dir in prot_dirs:
    prot = str(prot_dir).split("/")[-1]
    subdirs = sorted(os.listdir(os.path.join(prot_dir, "unsupp_cifs")))
    for subdir in subdirs:
        missense_df_path = os.path.join(prot_dir,"results/varalign/{}/{}_{}_missense_df.csv".format(subdir, prot, subdir))
        if os.path.isfile(missense_df_path):
            miss_df = pd.read_csv(missense_df_path)
            miss_df["protein"] = prot
            miss_df["group"] = subdir
            missense_dfs.append(miss_df)
        else:
            print("Group {} of {} did not present variation data".format(prot, subdir))
            pass
        
all_missense_df = pd.concat(missense_dfs).reset_index(drop = True)

all_missense_df.shape # total number of alignment columns with variation data

all_missense_df.head(3)

all_missense_df.to_pickle(os.path.join(results_dir, "all_miss_dfs.pkl")) # all variant-containing columns

## INTERACTIONS DATAFRAME

all_cons_dfs = []
for prot_dir in prot_dirs:
    prot = str(prot_dir).split("/")[-1]
    subdirs = sorted(os.listdir(os.path.join(prot_dir, "unsupp_cifs")))
    for subdir in subdirs:
        arpeggio_subdir = os.path.join(prot_dir, "results", "arpeggio", subdir)
        arpeggio_files = [file for file in os.listdir(arpeggio_subdir) if file.startswith("arpeggio_all_cons_split")]
        for file in arpeggio_files:
            pdb_id = file[-8:-4]
            file_df = pd.read_csv(os.path.join(arpeggio_subdir, file))
            file_df["struc"] = pdb_id
            file_df["protein"] = prot
            all_cons_dfs.append(file_df)
all_fragsys_cons = pd.concat(all_cons_dfs).reset_index(drop = True)

all_fragsys_cons.shape # total number of interactions

all_fragsys_cons.head(3)

all_fragsys_cons.to_pickle(os.path.join(results_dir, "all_bss_cons.pkl")) # all binding site contacts
