==> analysis/5_fragsys_feat_extraction.ipynb <==
# EXTRACTING UNIPROT FEATURE ANNOTATIONS TO CLASSIFY SITES BASED ON FUNCTION

In this notebook we automatically extract UniProt feature annotations via UniProt Proteins API.

### IMPORTING NECESSARY PACKAGES

from fragsys_analysis import *

### READING INPUT DATA

main_dir = "./../"
results_dir = os.path.join(main_dir, "results")

prots_df = pd.read_pickle(os.path.join(results_dir, "all_prots.pkl"))
all_bs_ress = pd.read_pickle(os.path.join(results_dir, "all_bs_ress.pkl"))

prots_df.head(5)

print(all_bs_ress.shape)

accs = prots_df.acc.unique().tolist()
print(len(accs))

acc_feat_dfs = []
for acc in accs:
    acc_feat_df = pd.read_json("https://www.ebi.ac.uk/proteins/api/features/{}?categories=DOMAINS_AND_SITES".format(acc))
    acc_feat_dfs.append(acc_feat_df)

feats_df = pd.concat(acc_feat_dfs).reset_index(drop = True)

feats_df.head(5)

feats_df.shape

dump_pickle(feats_df, os.path.join(results_dir, "up_func_anns.pkl"))

accs_in_df = feats_df.accession.unique().tolist()

[acc for acc in accs if acc not in accs_in_df] # not in data frame as it is unreviewed

feats_df = feats_df.join(feats_df.features.apply(lambda x: pd.Series(x))).drop(columns = "features")

cc = ["accession", "type", "description", "begin", "end", "molecule", "evidences", "ligand", "ligandPart"]

feats_df[cc].head(3)

feats_df.type.value_counts()

int_feats = ["BINDING", "SITE", "ACT_SITE"] # NOT: REGION, DOMAIN, MOTIF, REPEAT, ZN_FING, COILED, DNA_BIND

feats_df_filt = feats_df.query('type in @int_feats').copy().reset_index(drop = True)
feats_df_filt.type.value_counts()

feats_df_filt.end = feats_df_filt.end.astype(int)
feats_df_filt.begin = feats_df_filt.begin.astype(int)

## GETTING FUNCTIONAL RESIDUES FOR EACH PROTEIN

func_ress = {}
for acc, acc_df in feats_df_filt.groupby("accession"):
    func_ress[acc] = []
    for i, row in acc_df.iterrows():
        beg = row.begin
        end = row.end
        if beg == end:
            func_ress[acc].append(beg)
        else:
            func_ress[acc].extend(list(range(beg, end+1)))
            
func_ress = {k: sorted(list(set(v))) for k, v in func_ress.items()}

print(len(func_ress)) # 24/35 proteins have functional feature annotations

bs_cols = [
    "BS0", "BS1", "BS2", "BS3", "BS4",
    "BS5", "BS6", "BS7", "BS8", "BS9",
    "BS10", "BS11", "BS12", "BS13", "BS14",
    "BS15", "BS16", "BS17", "BS18", "BS19",
    "BS20", "BS21", "BS22", "BS23"
]

## CLASSIFYING EACH SITE AS KF (OVERLAP WITH FUNCTIONAL RESIDUES) OR UF

bs_ress_membership = {}
binding_site_func_dict = {}
for prot, prot_rows in all_bs_ress.groupby("protein"):
    for grp, grp_rows in prot_rows.groupby("group"):
        for bs_col in bs_cols:
            bs_id = "{}_{}_{}".format(prot, str(grp), bs_col)
            bs_rows = grp_rows[grp_rows[bs_col] == 1]
            if len(bs_rows) == 0:
                continue
            else:
                bs_ress = bs_rows.UniProt_ResNum.unique().tolist()
                bs_ress_membership[bs_id] = bs_ress
                if prot not in func_ress:
                    binding_site_func_dict[bs_id] = "UF"
                else:
                    intersect = set(bs_ress).intersection(set(func_ress[prot]))
                    if len(intersect) > 0:
                        binding_site_func_dict[bs_id] = "KF"
                    else:
                        binding_site_func_dict[bs_id] = "UF"


print(len({k: v for k, v in binding_site_func_dict.items() if v == "KF"})) # 29 KF sites

for k in list(binding_site_func_dict.keys())[:5]:
    print(k, binding_site_func_dict[k])

dump_pickle(binding_site_func_dict, os.path.join(results_dir, "prot_func_dict_auto.pkl"))
