==> analysis/6_manual_v_auto_feat_extraction.ipynb <==
# AUTOMATIC V MANUAL FUNCTIONAL ANNOTATION AND THEIR COMBINATION

In this notebook we compare automatically extracted function annotations from UniProt using the Proteins API with a manually curated annotation performed by JSU and combine both results.

### IMPORTING NECESSARY PACKAGES

from fragsys_analysis import *

main_dir = "./../"
results_dir = os.path.join(main_dir, "results")

bs_func_AUTO = load_pickle(os.path.join(results_dir, "prot_func_dict_auto.pkl"))
bs_func_MANUAL = load_pickle(os.path.join(results_dir, "prot_func_dict_manual.pkl"))

bs_func_MANUAL2 = {}
for k, v in bs_func_MANUAL.items():
    nk = "{}_{}_{}".format(k[0], k[1], k[2])
    if v == "non_active":
        nv = "UF"
    elif v == "active":
        nv = "KF"
        
    else:
        print(k)
        continue
    bs_func_MANUAL2[nk] = nv

fdf = pd.DataFrame(sorted(bs_func_AUTO.items()))

fdf[2] = fdf[0].map(bs_func_MANUAL2)

fdf.columns = ["bs_id", "auto", "manual"]

fdf.head(5)

len(fdf.query('auto != manual').sort_values(by = ["bs_id"]))

fdf.query('auto != manual').sort_values(by = ["bs_id"])

def f(row):
    if row['auto'] == "UF" and row['manual'] == "KF":
        val = "KF"
    else:
        val = row['auto']
    return val

fdf['comb'] = fdf.apply(f, axis=1)

fdf.auto.value_counts()

len(list(set([el.split("_")[0] for el in fdf.query('auto == "KF"').bs_id.unique().tolist()])))

fdf.manual.value_counts()

len(list(set([el.split("_")[0] for el in fdf.query('manual == "KF"').bs_id.unique().tolist()])))

fdf.comb.value_counts()

len(list(set([el.split("_")[0] for el in fdf.query('comb == "KF"').bs_id.unique().tolist()])))

dump_pickle(fdf, os.path.join(results_dir, "prot_func_dict_comb.pkl"))
