==> analysis/15_ML_predicting_rsa_labels.ipynb <==
# PREDICTING RSA CLUSTER LABELS FOR ANY SITE

## IMPORTING NECESSARY PACKAGES

from fragsys_ml import *

## FUNCTIONS

def generate_rsa_vectors(rsa_profs, max_len = None):
    """
    From an RSA profile dict, generates suitable RSA
    feature vectors to use as input for the MLP. If max_len
    is not provided, it will take the value of the longest site
    of the rsa_profs dict.
    """
    if max_len == None:
        max_len = max([len(v) for v in rsa_profs.values()])
    bs_vectors = []
    bs_vectors_dict = {}
    for bs_id, rsa_sig in rsa_profs.items():
        rsa_sig_len = len(rsa_sig)
        rsa_range_prop = [0 for i in range(10)] # now let us change to 10
        for rsa in rsa_sig:
            prop_i = int(rsa/10) # 10 RSA BINS: b1 = [0,10), b2 = [10, 20), ... b10 = [90, MAX)
            if prop_i > 9: # if greater than 100, put in 10th bin
                prop_i = 9
            #print(prop_i)
            rsa_range_prop[prop_i] += 1
        rsa_range_prop = [round(i/rsa_sig_len, 3) for i in rsa_range_prop]
        rsa_range_prop.insert(0, rsa_sig_len/max_len) # ADDING BINDING SITE SIZE RELATIVE TO MAX SITE SIZE (IN THIS CASE 40)
        bs_vectors.append(rsa_range_prop)
        bs_vectors_dict[bs_id] = rsa_range_prop

    vector_df = pd.DataFrame(bs_vectors, index = list(rsa_profs.keys())) # obtaining RSA vectors, which are the 11-element features used for the machine learning
    return vector_df

## INPUT DATA

main_dir = "./../"

results_dir = os.path.join(main_dir, "results")

rsa_profs = load_pickle(os.path.join(results_dir, "rsa_profs.pkl")) # using this as an example, but could be any {site_id: [site RSA profile], ... } dict

for i, (k, v) in enumerate(rsa_profs.items()):
    if i < 5:
        print("{} : {}\n".format(k, v))
    else:
        break

## GENERATING RSA FEATURE VECTORS DATAFRAME

vector_df = generate_rsa_vectors(rsa_profs)

X = vector_df.sort_index() # here, X is still a dataframe. Thus, we maintain binding site IDs

labels = X.index.tolist()

## RUNNING MODEL ON TEST DATA

model_path = "./../results/79619_2_model_epoch_49_train_acc_0.96.h5" # this is the name of the MLP model
final_model = keras.models.load_model(model_path)
final_preds = final_model.predict(x = X, batch_size = 27, verbose = 0)
rounded_predictions = np.argmax(final_preds, axis = -1)

## READ RSA CLUSTER LABELS

predicted_labels_dict = {labels[i]: rounded_predictions[i] for i in range(len(labels))}

for i, (k, v) in enumerate(predicted_labels_dict.items()):
    if i < 5:
        print("{} : {}\n".format(k, v))
    else:
        break
